refactor: Extract Transactionally to database package as helper function
