+++ internal/database/transaction.go
@@ -1,0 +1,40 @@
+package database
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+)
+
+// Transactionally executes a function within a database transaction. It commits the transaction
+// if the function succeeds, otherwise it rolls back. If rollback fails, both errors are returned.
+func Transactionally(ctx context.Context, db *sql.DB, f func(tx *sql.Tx) error) (err error) {
+	tx, err := db.BeginTx(ctx, nil)
+	if err != nil {
+		return err
+	}
+
+	defer func() {
+		if p := recover(); p != nil {
+			if rbErr := tx.Rollback(); rbErr != nil {
+				err = fmt.Errorf("panic occurred: %v, rollback error: %w", p, rbErr)
+			} else {
+				err = fmt.Errorf("panic occurred: %v", p)
+			}
+		}
+	}()
+
+	err = f(tx)
+	if err != nil {
+		if rbErr := tx.Rollback(); rbErr != nil {
+			return fmt.Errorf("transaction rollback error: %w, original error: %w", rbErr, err)
+		}
+		return err
+	}
+
+	if err = tx.Commit(); err != nil {
+		return err
+	}
+
+	return nil
+}
