+++ internal/item/repository.go
@@ -6,7 +6,6 @@ import (
 )
 
 type Repository interface {
-	Transactionally(ctx context.Context, f func(tx *sql.Tx) error) (err error)
 	CreateItem(ctx context.Context, tx *sql.Tx, item Item) (*Item, error)
 	GetItem(ctx context.Context, tx *sql.Tx, name string) (*Item, error)
 	UpdateItem(ctx context.Context, tx *sql.Tx, item Item) (*Item, error)
