+++ internal/item/repository/repository.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"database/sql"
 	"encoding/json"
-	"fmt"
 	"strings"
 	"time"
 
@@ -28,39 +27,6 @@ func NewRepository(db *sql.DB, errorHandler database.ErrorHandler) *ItemReposito
 		errorHandler: errorHandler,
 		queries:      *query.New(db),
 	}
-}
-
-// Transactionally executes a function within a database transaction. It commits the transaction
-// if the function succeeds, otherwise it rolls back. If rollback fails, both errors are returned.
-func (r *ItemRepository) Transactionally(ctx context.Context, f func(tx *sql.Tx) error) (err error) {
-	tx, err := r.db.BeginTx(ctx, nil)
-	if err != nil {
-		return err
-	}
-
-	defer func() {
-		if p := recover(); p != nil {
-			if rbErr := tx.Rollback(); rbErr != nil {
-				err = fmt.Errorf("panic occurred: %v, rollback error: %w", p, rbErr)
-			} else {
-				err = fmt.Errorf("panic occurred: %v", p)
-			}
-		}
-	}()
-
-	err = f(tx)
-	if err != nil {
-		if rbErr := tx.Rollback(); rbErr != nil {
-			return fmt.Errorf("transaction rollback error: %w, original error: %w", rbErr, err)
-		}
-		return err
-	}
-
-	if err = tx.Commit(); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 // CreateItem creates a new item in the database.
